Fix: 공격 범위 설정 오류 및 접속시 생성 오류

+++ Socket/TCPFighter_Server/TCPFighter_Server/Session.h
@@ -10,8 +10,8 @@ struct SESSION
     RingBuffer readQ;
     RingBuffer writeQ;
 
-    DWORD action; // 캐릭터 상태 (이동)
-    BYTE direction; // 캐릭터 시야 방향
+    BYTE action; // 캐릭터 상태 (이동)
+    BYTE direction; // 캐릭터 시야 방향 LL | RR
 
     short xPos;
     short yPos;
